refactor(llmobs): extract evaluation metric validation
